refactor(github): move github setup and issues under internal/github
Move-only: relocate githubissues/ and githubsetup/ to internal/github/issues
and internal/github/setup. They stay two packages - githubissues exports
Service/New and githubsetup exports Service/NewService/CommandRunner/RealRunner,
so a flat internal/github would collide on Service.

Package clauses now match their directories (issues, setup); every call site
already aliased these imports, the way agent/domain is imported as agentdomain,
so no call site reads differently.
