+++ internal/github/setup/service.go
@@ -1,7 +1,7 @@
 // Package githubsetup implements agentdomain.GitHubSetupService for the GitHub Action
 // CI setup flow. Every git/gh command runs through the injected CommandRunner
 // under a 30-second timeout so a wedged subprocess cannot hang the UI.
-package githubsetup
+package setup
 
 import (
 	"context"
