+++ internal/scheduler/githubscheduler/workflow_test.go
@@ -8,7 +8,7 @@ import (
 	scheddomain "github.com/inference-gateway/cli/internal/scheduler/domain"
 
 	config "github.com/inference-gateway/cli/config"
-	githubsetup "github.com/inference-gateway/cli/internal/services/githubsetup"
+	githubsetup "github.com/inference-gateway/cli/internal/github/setup"
 	yaml "gopkg.in/yaml.v3"
 )
